refactor: 한 컴포넌트에서만 사용하는 hooks 위치 변경
